+++ src/model/channel/mod.rs
@@ -164,7 +164,7 @@ impl Channel {
     /// In DMs (private channel) it will return None.
     #[inline]
     #[must_use]
-    pub const fn position(&self) -> Option<u32> {
+    pub const fn position(&self) -> Option<u16> {
         match self {
             Self::Guild(channel) => Some(channel.position),
             Self::Private(_) => None,
